+++ src/main/go/domain/divergence.go
@@ -3,8 +3,8 @@ package domain
 type PotentialDivergence struct {
 	HierarchyLevelKey          string
 	HierarchicalId             string
-	TotalMarketValue           int
-	TotalMarketValueDivergence int
+	TotalMarketValue           int64
+	TotalMarketValueDivergence int64
 	InternalDivergences        []*PotentialDivergence
 }
 
@@ -12,6 +12,6 @@ type DivergenceAnalysis struct {
 	PortfolioId               int
 	TimeFrameTag              TimeFrameTag
 	AllocationPlanId          int
-	PortfolioTotalMarketValue int
+	PortfolioTotalMarketValue int64
 	Root                      []*PotentialDivergence
 }
